Include (test)Namespace Dsl values in PsiPackage refactoring
We need a useScopeEnlarger to make the platform aware that the
PsiPackage might be used in Gradle Build Files as well as in sources.

We need a referenceContributor to test for particular Psi patterns for
applicability of our referenceProvider, whose only job is to get the
value of the string argument/value and feed that to
PackageReferenceSet, also providing an appropriate scope to tell the
resolver where to actually look for that package.

Groovy only because why would anything be easy?

Bug: 214570613
Test: manual
Change-Id: I6f56761a8a5e54f4cb817ebd681f4ef6705cba1d