Do not run slow operations on FileEditorProvider.accept
This changes the logic in the DesignerEditorProvider.accept so no slow
operations are required. We do a preliminary identification of the type
using the extension and resource folder without loading the PSI.
Once we get the request to create the editor, we identify the exact file
and run the expensive operation there.

Also, we remove the slow calls to identify the type from other paths
like the DesignFilesPreviewEditor.

Fixes: 286055058
Test: Existing
Change-Id: I8a0dfe642e25242946390bc355a668fc78c6ed6f