Offer Maven Import intention action based on resolved dependencies
As written, the import intention action is offered based on
"registered" (i.e. declared) dependencies in the project, and so would
be offered on symbols that resolve, albeit through transitive
dependencies.  This is probably not what is intended, given that to
the user this would be offering a do-nothing-useful intention action.

A more reasonable implementation of course requires the capability to
register dependencies, but should query whether to offer the action
based on resolved dependencies, effectively asking whether the known
symbol currently resolves in the project.  Change `dependsOn` to do
that.

Although this looks like a change in current behaviour, it in fact
isn't because the Gradle Module System has disabled support for
inquiring about registered dependencies for several years, because the
current implementation requires parsing the project in an
EDT-incompatible way, instead delegating to the equivalent query on
resolved dependencies.  Forthcoming changes change the semantics of
getRegisteredDependencies, but in a way that would make this intention
action behave in a non-ideal way.

Bug: 129297171
Bug:
Test: existing
Change-Id: I8d48451b7bb2c84384e315584897c87f7c4acfab