Improve BuildInvoker capablities
The capabilities of a build invoker are used like the hard requirements
by the callers to obtain an appropriate invoker. For example, if a
caller needs to build an android instrumentation test, it calls
getBuildInvoker with a required capability BUILDS_AIT.

This change updates the capabilities in such a way that they represent
what caller requires, instead of what a build system can do.

There is no behavior change in the IDE, the following changes are made
to the capabilities based on what they're used for.
1. IS_LOCAL -> BUILD_AIT.
2. SUPPORTS_PARALLELISM -> BUILD_PARALLEL_SHARDS.
3. SUPPORTS_API -> RUN_REMOTE_QUERIES.
4. SUPPORTS_DBIP -> removed.
5. DEBUG_LOCAL_TEST -> added.

Bug: 374906681
Test: n/a

Change-Id: Id29c1e5eb6e3561552a8edc1e334963da97e2da3