Return NO_STATUS test result if no tests were run
The BES does not send a TEST_RESULT event if the tests were not run (for
reasons such as build failures). Consequently, ASwB displays an
incorrect `All tests passed` message to the user.

This change isn't ideal, but works towards fixing this by adding a
`NO_STATUS` test result while parsing the BEP.

Bug: 400687547
Test: n/a

Change-Id: Ie2b694506515e80df6b3bcc4d33b1f3d5e7f8a5e