containingSymbol -> containingDeclaration
In the case of top-level symbols, querying for the containing symbol
would end up resolving a KaFileSymbol, which is unnecessary in
this case since we are only interested in symbols nested inside R
classes. Furthermore, for unknown reasons, KaFileSymbol resolution
sometimes throws an exception when in K1 mode with IntelliJ 2025.1.

We can use containingDeclaration instead.

Fixes: 401071045
Test: AndroidLintTest.testAddSdkIntKotlin() + IJ 2025.1 + K1 mode
Change-Id: I4e2e51dc4648a7003a4b2f1f465c134939cbe28d