Move disposed handling to SceneManager
Before, isDisposed was controlled at the LayoutlibSceneManager level.
Because the actual Disposer registration happens in the SceneManager and
we register during the constructor, it means that there could be cases
where the dispose call happens before isDisposed had the chance to be
initialized to its initial value.

Moving it to the SceneManager constructor will fix this issue.

Test: Existing
Fixes: 400416804
Change-Id: I65baf669248e7410d2a0f91c47553fcf2b464e29