Update devices directly without using updateDevicesChannel
When investigating the linked bug, it showed that when unpairing two
wear devices from a phone only one was unpaired on windows. In reality,
the second was also unpaired, but with a 60s delay.

It would appear that the updateDevicesChannel.trySend would get stuck
until the updateDevicesChannel.receive() would timeout (after 60s) and
then it would resume removing the next wear device.

This CL replaces the use of the updateDevicesChannel with direct calls
to updateListAndForwardState.

This CL also adds a try/catch to updateListAndForwardState.

Fixes: 332844801
Test: existing + manually tested
Change-Id: I1697419aa60bcd509e21d17cf673044b805de348