compose-designer: remove stubDevicesAsLibrary()
This line leads to a duplicate class definition for
androidx.compose.ui.tooling.preview.Devices (at least in
testGroupValuesProvider), which confuses Kotlin resolution and
triggers an assertion failure (when running with IntelliJ 2025.1).
This is a bug in the Kotlin IDE plugin, but we can work around it
in the meantime.

Other test cases appear to pass without this line.

Bug: 401299108
Test: PreviewPickerValuesProviderTest
Change-Id: I238e02032264972c99359fd1fed2fbfdae839979