Do not leak ViewInfo objects in VisualLintRenderIssue
Visual Lint issues are attached to the Problems tool window and can
survive beyond the lifetime of the render model that they are based off.
They should not reference any ViewInfo objects as that unnecessarily
holds on to memory.
This ensures that no references to ViewInfo objects exist in the issue
description providers.

Bug: N/A
Test: N/A
Change-Id: I69d6d4b2b80bdc1495a5bad1c1aa74c37a89e308