Fix flakiness around AndroidDomElement
Flaky tests after the 2025.1 merge showed that
AndroidDomElementDescriptorProvider seemingly has a race condition with
DomManagerImpl.

This fix brings AndroidDomElementDescriptorProvider in line with the
logic in the platform's DomDescriptorProvider. I'm not sure whether this
is a real fix or just masking the underlying race condition, but since
it empirically allows the tests to pass and mirrors the platform logic,
I'm okay with adding it to fix this issue in the short term.

I will also follow up on the associated youtrack issue (linked below) to
try to determine if there are platform changes necessary here.

YouTrack: https://youtrack.jetbrains.com/issue/IJPL-180788

Fixes: 402201770
Test: Multiple (included)
Change-Id: I0fc1c564b86644aedd5b8525082c74a1a0db431a