Rework getting a library display name
The artifact address, if filled in by the Gradle sync model factories,
is (usually) of the form "group:name:version@suffix" for suffix being
`aar` or `jar` (some other forms for "local aars" or "local jars" are
also possible).

For those in the form above, the part before the suffix is a (Gradle)
Component identifier.  Use that to parse it and hence generate the
display name for the manifest panel.

Bug: 279886738
Test: existing
Change-Id: I829bfd54e311604aedffd3343b94f702b86dcba1