Move analysis out of write actions
Modifying or deleting Psi parameters triggered an analysis within
PsiCallParameterPropertyItem.value. This analysis was being invoked
within a write action (specifically, inside the runModification
block), leading to a ProhibitedAnalysisException.

To resolve this, the analysis actions within
PsiCallParameterPropertyItem.value have been moved outside the
runModification block. This ensures that the analysis is performed
outside of the write action context, preventing the
ProhibitedAnalysisException.

Fixes: 404173904
Fixes: 404266894
Test: PreviewPickerTests
Change-Id: I4f5f655ac726d6c62c97aae7159b7a0277ab2db4