Return declared class names in internal format
The function getDeclaredClassNames was returning class names in
com.example.MyClass rather than com/example/MyClass, which is the format
the differ uses. This caused Live Edit to treat every class as a
compiler-generated class, and thus as a proxyable class.

Existing tests didn't catch this because the format difference is only
apparent when examining classes outside of the default package; a test
case has been added for this.

Fixes: 402573127
Test: ComposableCompileTest#`File in non-default package`
Change-Id: I7b432ad86c72e46ab92eb70b736f56dc5531b0a5