Add a drawable resource annotator for Declarative Watch
... Faces.

Declarative Watch Faces can have attributes that reference drawables.
When these drawables are referenced with the "@drawable" prefix they are
properly annotated. However, DWF resources are not required to be
prefixed with "@drawable"[1]. This CL adds an annotator that handles
annotates drawables referenced from the "resource" and "icon" attributes
in DWF XML files (in res/raw).

[1]: https://developer.android.com/reference/wear-os/wff/group/part/
image/image?version=3#required-attributes

Bug: 392849349
Test: DrawableResourceExternalAnnotatorTest
Change-Id: If27bd850f03d47c2622a9fdfb0ab89d129633887