Store related dagger items with different keys
The error in b/405380916 indicates that a single PSI element is
identified both as an `AssistedInject` constructor and as a `Inject`
constructor simultaneously. This isn't a configuration that will build,
but it can theoretically happen in the editor if the dev adds both
attributes.

By using a different cached value key for each type of dagger element,
we can avoid collisions in corner cases like this.

I haven't been able to directly reproduce this unfortunately, and
therefore haven't found a way to trigger this and test it.

Fixes: 405380916
Test: n/a
Change-Id: I039428d0a1b2ff7747e79db61858681e26d495a3