Fix flakiness in PsiUtilsTest
The utility methods being tested require being called when not in dumb
mode. Wherever they're used in the product, something higher up in the
stack makes that guarantee; but nothing in the tests does.

Recent platform changes have made this issue come to light, so I'm
fixing this by ensuring all the read actions calling into these utility
functions first wait for indexing to finish. This is a little bit of
"killing a fly with a sledgehammer", but it should help address what
amounts to a whole bunch of super-rare flakes that can add up to much
more flakiness overall.

Fixes: 405423790
Presubmit-Runs-Per-Test:
//tools/adt/idea/intellij.android.compose-common:intellij.android.compose-common.tests_tests@200
Test: PsiUtilsTest
Change-Id: I18a872bd145bf1869a5a5a990098ff4694f39c62

Change-Id: I7f8d085f48760845c5a58c5d793c1cd6b53e1bc5