Move client handling logic out of RootPanelRenderer
OnDeviceRenderer panel holds an instance of OnDeviceRenderingClient,
which goes stale each time the client is disconnected. For this reason
it is necessary to recreate OnDeviceRendererPanel each time a new
connection is established.

This change is a refactoring of how OnDeviceRenderingPanel works.
Now the panel that does the actual on-device rendering is private and
can only be instantiated by a new OnDeviceRendererPanel class that acts
as a manager for it, making sure that the current instance is never
connected to a stale client.

Doing this hides away the implementation details of using a client and
simplifies how clients can use on device rendering.

Test: OnDeviceRendererPanelImplTest
Test: OnDeviceRendererPnalTest
Test: RootRendererPanelTest
Bug: 403288855
Change-Id: I8066135106e65fc45e7fbb9e1623791e78bf2235