Attempt to deflake test
Not entirely sure what's going on here. The stack trace doesn't look
like it's directly related to the test. It seems to be related to the
underlying project disposal. This change attempts to address any
potential race conditions by putting the project rule on edt.

As of now, the test flakes about once a week, so definitely not
terrible.

Bug: 405361416
Test: this is it.
Change-Id: I3a4faceae0d8f6b86f3aefb62f011ab1e8b9461b