Remove need for spurious empty build.gradle.dcl for tests
It was included because the change in strategy for detecting which Psi
changes are relevant for sync banner state in the previous change
causes events associated with files created since the last sync to be
considered as irrelevant.  This changes behaviour at the margins but
is unavoidable with the current strategy.

Adjust the test to modify an already-existing declarative gradle build
file for now.  The proper fix for all this will be to move to the
IntelliJ platform's ExternalSystemProjectNotificationAware machinery
which handles tracking VFS as well as Psi events for project-relevant
files.

Bug: 399868758
Bug: 406493217
Test: test changes only
Change-Id: I2a9c02f45277bed8369d81a9bd796354675087e9