Add capability to fetch a device being debugged
Some of the Android Studio debugger features require fetching a device
that is currently being debugged. For example, when performing
smart step target filtering, the debugger needs to fetch an APK from
the client, see b/396095270. Also, when loading debugger helper classes
to a client VM, the debugger needs to now which version of dex bytecode
to produce. See b/403260091.

Test: N/A
Bug: N/A
Change-Id: I11e10298dd40e8ceb0529d750acba48c9e66282f