Set minimum width for text fields in the string resource panel
It turns out that a JTextField cannot compute a reasonable minimum
width for Bangla text, instead the actual width is returned.
This is causes the parent panel to clip the text such that the only
the first part of the text can be edited.

Set an arbitrary width of a text field which allows the text field
to scroll horizontally such that the entire string can be edited.

Bug: 402308111
Test: StringResourceViewPanelTest
Change-Id: I6207c74bc476d8a2bbc685d0bd662e694ff3959c