Clear Module class loader data when clearing cache
When clearing cache associated to a module in
StudioModuleClassLoaderManager, we only clear the module data from
modules referenced in holders. That is an issue when the cache clearing
happens when no module is in holders, e.g. when resyncing the project
without a preview showing. In this case, the module would keep outdated
data.
This ensures that we always clear the module data, whether the module is
in holders or not.

Fixes: 400375502
Test: N/A
Change-Id: I6d58eb61b69c94ad267e3b8c2e3233e29046b743