APK Analyzer UX feedback updates for 16 KB Page
o Merge Zip Alignment and LOAD Alignment into a single Alignment column.
  - If there is an alignment problem, then show a warning describing it.
    Example, "4 KB, but 16 KB is required".
  - Else, if both alignments are the same, then show that value.
    Example, "16 KB".
  - Else, if alignments are different, then show them both.
    Example, "16 KB zip|4 KB LOAD section".
o Default open the nodes that have warning so the user can see them
  easily.
o Clicking on an unaligned .so will show a message with a link to docs.
o Optimize use of horizontal space.
  - Reduce some column widths.
  - Move the Alignment column all the way right.
  - Rename "Compression" column to "Compressed" and make the column
    values "yes" and "no" rather than "Compressed" and "Uncompressed".

See go/as-16kb-bugbash for screenshot of APK Viewer.

Bug: 382091362
Test: Unittests added
Change-Id: I7cc919eed9c90d1085813e48b9138f60f4447dff