Creates kotlin facet for query sync project
Just create with default values. Will update the plugin class jar path
in next cl.

It's nearly the same as BlazeKotlinSyncPlugin. Ideally, we should not
keep duplicate code blocks but it's intend to keep separate
implementations for different sync mode. So do not merge them until we
think it's necessary.

Test: existings tests
Bug: b/406257186
Change-Id: I68b9efeaa517931bef779f3a112fc40dd1445588