Deprecate a suboptimal interface
Getting a MessageBus connection using the project as a Disposable is
potentially leaky.  Deprecate the method that does this implicitly, in
favour of methods which explicitly specify the disposable.

This change deprecates the error-prone interface and adjusts uses in
tests.  There remain some production uses which will need further
consideration.

Bug: 408981671
Test: existing tests adjusted
Change-Id: Iaa0ad443a3e6a2716afada2c49fd110a227f9a28