Deprecate AndroidCoroutineScope, improve other APIs
Given that AndroidCoroutineScope uses a deprecated API in its
implementation, and we have createCoroutineScope() which doesn't,
this CL deprecates the former in favor of the latter.

It also adds the exception handler to createCoroutineScope() to
maintain the existing behaviour from AndroidCoroutineScope, although it
might not be necessary — I am unclear on this.

Bug: none
Test: none, there is no functional change not already covered
Change-Id: Ic272c6f5fb75f14c3971736e3fcef47b9ddd4e54