Ignore module file for variant switcher snapshot tests ...
... for phased sync.

Variant switching works by storing and restoring the data nodes by
running them through the data services again when switching back to a
known variant. This is basically bypassing anything done in via phased
sync (meaning the module file representation will be back to being a
.iml based one), so we have to ignore the module file in these tests
until there is an actual solution to this.

To be clear this is also relevant in production but the effect of this
is not disastrous, at least not that we are not aware of (just the
module file representation changes).

Bug: 384022658
Test: self
Change-Id: I96e4c669e831cfe0e7bfd6854f2bc6ded5129920