Tighten a utility
Implicitly the one meaningful production caller of
Project.getModuleByGradlePath() assumed that it was getting the holder
module (specifically the module whose scope encompasses proguard
files).  Rename the utility to make this explicit, and likewise
perform the explicit test when filtering for the module to avoid
accidentally returning a sourceSet module.

Bug: none filed
Test: existing, exposed by phased sync experiments
Change-Id: Ie17bc3d906c209254aded13fc228ac17435c2c3f