Fix wrong auto import in preview picker
When selecting a UI Mode, new parameters might require importing
classes (e.g., Configuration). While imports are usually added with a
newline, this doesn't seem to be happening for Configuration.

The cause of this problem was because once we selected a uiMode, a
functionality that shortens the imports named ShortenReferenceFacility
is called. To perform the shortening of the namespace we were using
runUndoTransparentAction while instead writeCommandAction was needed.

A write command action is needed when we are performing a text
replacement actions like ShortenReferenceFacility.

Fixes: 407880030
Test: PreviewPickerTests
Change-Id: I146e74e82f0c833a915c4d06c3e3abd26fe275b1