Reduce memory usage when checking if we need to create a
... Declarative Watch Face run configuration when the watch feature is
not required.

Calling application.activities uses a ResourceRepository. If the watch
feature is not required in the manifest, we don't need to check the
existence of activities or services.

This CL changes the logic of the createDeclarativeWatchFaceConfiguration
early return if the watch feature is not required. We only check the
manifest's activities and services when the watch feature is required.

Bug: 410510784
Test: existing tests should still pass, ran
 //tools/adt/idea/profilers-integration:ProfileWithCompleteDataTest to
 check that the memory usage was back to normal
Change-Id: Id324b4cf9dda58a922826d178c620499f7f89df1