Clean up hack used to decide if a parameter is optional or not
For `fun KaSession.isRequired(valueParamSymbol:
KaValueParameterSymbol)`, we checked whether the value parameter ends
with "/* = compiled code */" or not to decide that it is a parameter
with a default value. It was a hack for a situation that it has a
default value but the value parameter symbol `hasDefault` does not work.
When we tested on an Android Studio instance with K2 for multiple
compose samples like JetNews, we did not meet a case that the default
value decision can be done with the hack. This commit cleans it up.

Bug: 274145999
Test: Manual check on Android Studio instance
Change-Id: I19797837aa58597a95d7235fb5092e5be5d610e0