Use TestDialogManager to avoid flakiness
Because tests in AndroidGradleProjectStartupActivityTest
interact with code that results in showing a dialog, this
should be accompanied by TestDialogManager call (setting a
reaction to those dialogs shown in test mode). Otherwise the
RuntimeException has been witnessed. This is rare, as might
have been influenced by factors set by preceding tests.

Bug: 410561122
Test: Existing
Change-Id: I38f20a732ea20e55eb475b930cc1bf1f7272ce34