Make query sync action execution flows more linear
i.e. make the flow begin and end in
`BuildDependenciesHelper.determineTargetsAndRun`.

Build and sync like actions are expected to be invoked in the EDT, where
they initialize the UI and launch build/sync. The build itself runs in
the background. Some of these actions may return a future of build
results.

1. Make all actions return a future;
2. Change `determineTargetsAndRun` to accept a function returning a
   `Deferred` result of the build action;
3. Make target disambiguation popup notify when it is cancelled and
   use the callbacks to complete a future only.
4. Express the flow in a way that starts and finishes in the same
   function and not in a callback.

Bug: 409737053
Test: n/a
Change-Id: Idb54f65bcb195d9741aea12e01ce71257d126775