Ensure project SDK uses resolved Daemon toolchain
Until now, the IDE was responsible for indicating to Gradle
which toolchain to use after resolving the Gradle JDK configuration,
using the same to create the project SDK, if required. However, this
will no longer be the case when using Daemon toolchain, since Gradle
internally will need to find based on the defined criteria the
installed toolchain or download a compatible one.

For this reason, IDE will use the IdeaProject model to obtain the
resolved JDK based on the criteria used to run the build and align
project SDK with it.

Bug: 395610076
Test: SingleGradleRootSyncUseDaemonJvmCriteriaIntegrationTest
Change-Id: I7df156d43a23df02ce5a7d7173d15752dbdd293e