Fix scope handling for running lint checks with single-file scopes
Some lint checks use different code paths when we're running in isolated
mode (single file analysis, e.g. running on the fly in the editor)
versus running in "batch" mode.

The way this is determined is whether the lint scope is just a single
file type. However, users can also, when invoking lint manually, pick
the current file as a scope. In that case, lint checks will pick up the
isolated mode behavior, but in this code path we were not also doing the
other thing lint does for on the fly analysis: set up a "main" project
which points to things like the app minSdkVersion etc. This affected
things like the 16 KB alignment check, which needs to see the dependency
graph.

Fixes: 410936202
Test: Existing + new test
Change-Id: I826fd1b210c0f529ad28eba79554f0c248a995a0