Update logic in "RepositoryUrlManager" to pass in "GoogleMavenRepositoryV2Host".
This changes makes changes to construction logic of "RepositoryUrlManager" to pass in the newly created "GoogleMavenRepositoryV2Host" to "GoogleMavenRepositoryV2". Addtionally, it updates the unit tests to use the "FakeGoogleMavenRepositoryV2Host". The "ENABLE_GMAVEN_REPOSITORY_V2" flag is still false so it doesn't result in any logical changes for the callers of "RepositoryUrlManager".

Bug: b/407727269
Test: Unit tests
Change-Id: Ic971075041fd6a56c8cd6d3afcfeca8710acb9a6