Add parser for multiple build failures
This introduces parser for gradle multi-failure message, which is based
on original GradleBuildScriptErrorParser with some refactorings
applied for simplification.

The failure message structure is well-defined so we first parse all
reported failure messages and then process each of them the same way
as in GradleBuildScriptErrorParser.

Additional hack has to be introduced in BuildOutputParserManager to
mitigate the problem in platform's GradleBuildScriptErrorParser. It
does not expect this multi failure message as a build completion line
and does not reset current output reader resulting in wrong ids set for
the generated events.

This change covers most of the cases though several todos are left for
the follow-up CLs:
 - We have previously introduced several specified parsers that also
 read build failure message. They as well do not expect multi-failure
 message at the moment and need to be adapted.
 - Original GradleBuildScriptErrorParser invokes list of IssueCheckers
 to generate additional specified messages or skip them if they are
 already reported from analyzing exceptions directly. This can not be
 adopted as is at the moment as most of the issue checkers are not
 adapted to analyze multi-failure exceptions. This means in this case
 they are likely not to report the message from exception and it is
 better not to filter such messages out. Risk of duplication in this
 case is better than risk of not showing any error at all.

Bug: b/362959090
Test: GradleMultipleFailureOutputParserTest, all other Build Output
  Parser tests verify system works as expected altogether.
Change-Id: I547299ec33cf2e0386f7f2d7fd91fb6a33cf6649