Remove duplicate jars from compile_jar list of target
We notice the duplication jars due to java_info.compile_jars may return
dependencies' jar in some cases. To fix it, in aspect we saved the java_output.compile_jar, and expose that information. In IDE, we can generate a map from jar to the target generate it. Then we will not add jar to library if it's generated by other target but list in target's compile jars list.

Bug: b/410861643
Test: SyncedInBazelProjectTest
Change-Id: Ibfc7e576097938e8b47bab6609ca8736addaeff3