[settings sync] Pick up google login state changes
...and react on login/logout.

On login, we 1) clear login action installed if there's.
2) perform another sync to refresh sync status.

On logout, we perform another sync and we install a
login action for users to log in on the B&S page.

Note, this is a compromised change to accommodate the
current platform design. A cleaner way could be done
later when we have the upstream merge where we have
available disposable.

Also, there's missing "requestSuccessful" implementation
included in this change.

Bug: 411744911
Test: added
Change-Id: I93becbe08e40d90d3a98461c444f2cabe2a2e306