Reduce scrolling speed when using touchpad
The  WheelEvent proto using integers for dx and dy as opposed to floats used by Android. Even when a touchpad scroll event is mapped to WheelEvent with dy = 1, the speed is too high. To achieve a reasonable speed, small scroll events are being accumulated and produce a WheelEvent when the cumulative scroll amount crosses an integer boundary.

Bug: 411486773
Test: EmulatorViewTest
Change-Id: I91ee373ecf0475175ba5fa152f52419509879bff