Give Deployment the Ability to Terminate
The deployment process on newer Android versions (API 33+) can start to
rely on the package manager to terminate an application before
installation.

However, in scenarios like install-without-install (IWI), the
deployment logic may need to explicitly terminate the app.

This change introduces a `DeployerApplicationTerminator` to provide the
deployment flow with a formal mechanism for application termination.

Bug: 412689373
Test: Existing IWI ETE test should pick up the not killing issue (when we toggle the flag on)

Change-Id: I603ba9a69b05359b715536c00c52b54b9f6b4ae4