Remove no-effect UI customization
It might have worked once to find ActionButtons related to specific
text presentations to remove them from the UI, but with the current
implementation of ToolbarDecorator, any buttons are created (a) later
than the dialog initialization and (b) possibly in a separate
component tree.  Attempting to remove the Exclude button this way has
not worked since at least Arctic Fox, and no-one has complained, so
just remove it.

Bug: none filed
Test: not covered by tests
Change-Id: I1d001a74436647e88b511d710e1e0d43ba2dc8cc