[Dsl] Allow deleting useLibrary
This completes a minimal set of operations for `useLibrary`, since
users can now examine the useLibrary calls that are present, and take
whatever action they require (including deleting and re-creating with
different options if necessary).  Improving the ergonomics of this
support, or integrating better with the rest of the Dsl model, is left
to be done when it is needed.

Bug: 410807735
Test: added
Change-Id: I3d4c68137a886de413e5308edcd59998c01e2a89