Add phased sync consistency test
These check the intermediate state achieved by executing phased sync to
the full sync state. These tests are run for many of the snapshot
projects.

There are many known issues encoded within the tests and we need to
solve these one-by-one. So these tests also serve as an indicator of
known issues and how to proceed.

To implement these tests I've modified the dumper to output a different
format to denote the hierarchy and names for each line. These lines are
then filtered/compared across the intermediate and full state. Example:

MODULE (gradle_project.app)/FACET (Android)/SelectedBuildVariant (debug)

I also had to add an "ignore tasks" toggle in the dumper separately
because it was triggering a value cache for the project path too early
and causing issues.

Bug: 384022658
Test: Adding tests
Change-Id: I0950dffdc513266e09eb13d6f3277d2a249638a2