Store the shard count used by legacy full sync in import settings
This will be used later to decide if the project may be auto converted. The value is stored in `BlazeImportSettings` and will be persisted in `.idea` folder of the project. QuerySync auto-conversion will be run on project re-open and it would read this value from the persisted `BlazeImportSettings` in `.idea` folder of the project.

Bug: 405438522
Test: Manually verified that stored value is retrieved on project load.
Change-Id: If2063411ec4ab8e6311f15c4d5b9fad3f8c2c70e