Force KaModules for Android light Java classes.
The new K2IDEProjectStructureProvider no longer uses IDEA's ModuleInfo
classes to build project structure, instead using a new implementation
based directly on the IDE's workspace entities. This means that the
overrides put in place to ensure that Android light classes, such as
R classes, are associated with the correct module no longer function,
since those attributes are not read by the new PSP. This results in
Android light classes being registered as out-of-content-root classes,
which leads to issues when trying to retrieve their FIR symbols.

To work around this issue, we implement the (internal-only)
K2KaModuleFactory EP to force the KaModule created by the new PSP to be
the same module as the light class's associated module, library, or SDK.
This will serve to work around the issue until a more general solution
for overriding KaModules can be upstreamed.

Bug: 415168694
Fixes: 412606827
Test: Included
Change-Id: Ic40a3d9e77ade6935a678b18c3eaf1ea2b0d0d95