Improve import project with defined .gradle/config.properties file
When opening a project with .idea directory this is considered already
imported, resulting on GradleProjectImporter not being triggered. This
expected behaviour results in the case of .idea/gradle.xml file not
being present to use by default the #USE_PROJECT_JDK gradleJvm.

For this reason studio always considered mandatory to have added
.idea/gradle.xml on VCS to indicate how to resolve current
Gradle JDK configuration. Since this created some confusion to users,
now if project contains .gradle/config.properties file
pointing to a valid java home, the gradleJvm will be configured
with #GRADLE_LOCAL_JAVA_HOME macro resulting on the specified path
to be used to set up Daemon JVM for build execution.

Bug: 414991402
Test: GradleJdkConfigurationRecoveryIntegrationTest
Change-Id: I45f256dd8e3c69104064114bf5ce940b37224b58