Add java aspect test
This is a set of test cases that we want to verify if
build_dependencies_deps.bzl file works as expected. This file is used to
retrieve different provider for different language. And we need to
verify its output contains the expected information e.g. some jar.

To verify it, we provide a build_dependencies_deps_test_fixture that
will use build_dependencies_deps as aspect for each test target and
collect the output we nned to test and return to the them. In each
test case, they will be checked according to the rule and language
(e.g. java_library, java_binary).

These test cases will be copied to google3, so that both blaze/ bazel
will be tested.

Bug: b/377865954
Test: added
Change-Id: I35687267aa80514c8a1f82fdf3e91a26323ee36c