Create AndroidGradleSourceSetEntitySource
The generated files are generated via the "Generate workspace model
implementation" action in IDEA.

Just to note, I had to disable K2 in my IDEA to make it work after
seeing some related errors in the IDE when running the action. This
just means the action doesn't work. The generated work is still fine
in K2 after generating in K1 mode.

Bug: 384022658
Test: N/A
Change-Id: I55b19ad26a038233b4bccec60c02af16259762f5